Add `--no_nc_ss_adjustment` argument
This argument bypasses the default heuristic
check for putatively-misannotated U12-type
splice sites in introns with non-canonical termini. In
other words, by passing this argument all intron annotations
will be used as-is. Usually, putative misannotations comprise only a
small number of total introns, but this option ensures the output
matches the input exactly.